requirements.txt
# NAOMI Agent Dependencies
pyyaml>=6.0
aiohttp>=3.9.0
aiofiles>=23.0
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
websockets>=12.0
duckduckgo-search>=4.0
psutil>=5.9.0
watchdog>=3.0.0
httpx>=0.25.0
python-multipart>=0.0.6
anthropic>=0.90.0
beautifulsoup4>=4.12.0
# v1.3: TTS
edge-tts>=6.1.0
# v1.5: Local MLX TTS (Apple Silicon)
mlx-audio>=0.1.0
soundfile>=0.12.0
# v1.3: Gmail & Calendar
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
# v2.0: Intelligence upgrade
sqlite-vec>=0.1.6
sentence-transformers>=3.0.0
playwright>=1.40
